Project

+++ src/main/java/com/belyuk/first_project/entity/IdGenerator.java
@@ -1,11 +1,11 @@
 package com.belyuk.first_project.entity;
 
 public class IdGenerator {
-  private static long iD = 0;
+  private static int iD = 0;
 
   private IdGenerator() {}
 
-  public static long getIdCount() {
+  public static int getIdCount() {
     return ++iD;
   }
 }
